+++ tests/factories.py
@@ -232,6 +232,7 @@ class VectorSettingsFactory(BaseFactory):
         model = VectorSettings
 
     name = factory.Sequence(lambda n: f"vector-{n}")
+    title = "Vector Title"
     dimensions = factory.LazyAttribute(lambda _: random.randrange(16, 1024))
     dataset = factory.SubFactory(DatasetFactory)